+++ lua/gitlab/ui/picker.lua
@@ -49,10 +49,6 @@ function M.show_pipeline(opts)
   return load_backend().show_pipeline(opts)
 end
 
-function M.show_jobs(opts)
-  return load_backend().show_jobs(opts)
-end
-
 function M.show_job(opts)
   return load_backend().show_job(opts)
 end